src/frontend/requirements.txt
# Add requirements here - ref backend/requirements.txt
streamlit
requests
black
protobuf
altair
pillow
